requirements/base.txt
django==4.1.4
django-environ==0.9.0
whitenoise==6.2.0
Pillow==9.3.0
djangorestframework==3.14.0
psycopg2==2.9.5